backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "pypdf2==3.0.1",
    "reportlab==4.0.7",
    "pikepdf>=8.7.1",
    "pypdf==3.17.1",
    "pillow>=10.1.0",
    "pymupdf>=1.24.0",
]

